requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9.0
einops>=0.6.0
tensorboard>=2.13
pyyaml>=6.0
